measurementpoint crud

+++ resources/js/measurement_point.js
@@ -2,6 +2,118 @@ var inputprojectId = null;
 var userList = [];
 var modalType = "";
 var inputUserId = null;
+var inputMeasurementPointId = null;
+var noise_meter_data = [];
+var concentrator_data = [];
+
+function populateConcentrator() {
+    var selectConcentrator;
+    const defaultOption = document.createElement("option");
+
+    if (modalType === "update") {
+        selectConcentrator = document.getElementById(
+            "selectUpdateConcentrator"
+        );
+        selectConcentrator.innerHTML = "";
+        const defaultConcentrator = concentrator_data[0];
+        defaultOption.value = defaultConcentrator.noise_meter_id;
+        defaultOption.textContent = `${defaultConcentrator.device_id} | ${defaultConcentrator.concentrator_label}`;
+    } else {
+        selectConcentrator = document.getElementById("selectConcentrator");
+        selectConcentrator.innerHTML = "";
+        defaultOption.value = "";
+        defaultOption.textContent = "Select Concentrator...";
+        defaultOption.disabled = true;
+    }
+    defaultOption.selected = true;
+    selectConcentrator.appendChild(defaultOption);
+
+    const url = "http://localhost:8000/concentrators/available";
+    fetch(url)
+        .then((response) => {
+            if (!response.ok) {
+                throw new Error(
+                    "Network response was not ok " + response.statusText
+                );
+            }
+            return response.json();
+        })
+        .then((data) => {
+            console.log(data);
+            data = data.concentrators;
+
+            // Create options from fetched data
+            data.forEach((concentrator) => {
+                const option = document.createElement("option");
+                option.value = concentrator.id;
+                option.textContent =
+                    concentrator.device_id +
+                    " | " +
+                    concentrator.concentrator_label;
+                selectConcentrator.appendChild(option);
+            });
+        })
+        .catch((error) => {
+            console.error("Error fetching data:", error);
+        });
+}
+
+function populateNoiseMeter() {
+    const defaultOption = document.createElement("option");
+
+    if (modalType == "update") {
+        var selectNoiseMeter = document.getElementById(
+            "selectUpdateNoiseMeter"
+        );
+        selectNoiseMeter.innerHTML = "";
+        const defaultNoiseMeter = noise_meter_data[0];
+
+        defaultOption.value = defaultNoiseMeter.concentrator_id;
+        defaultOption.textContent = `${defaultNoiseMeter.serial_number} | ${defaultNoiseMeter.noise_meter_label}`;
+    } else {
+        var selectNoiseMeter = document.getElementById("selectNoiseMeter");
+        selectNoiseMeter.innerHTML = "";
+        defaultOption.disabled = true;
+        defaultOption.value = "";
+        defaultOption.textContent = "Select Noise Meter...";
+    }
+    defaultOption.selected = true;
+    selectNoiseMeter.appendChild(defaultOption);
+
+    const url = "http://localhost:8000/noise_meters/available";
+    fetch(url)
+        .then((response) => {
+            if (!response.ok) {
+                throw new Error(
+                    "Network response was not ok " + response.statusText
+                );
+            }
+            return response.json();
+        })
+        .then((data) => {
+            data = data.noise_meter;
+
+            data.forEach((noise_meter) => {
+                console.log("noise_meter");
+                console.log(noise_meter);
+                const option = document.createElement("option");
+                option.value = noise_meter.id;
+                option.textContent =
+                    noise_meter.serial_number +
+                    " | " +
+                    noise_meter.noise_meter_label;
+                selectNoiseMeter.appendChild(option);
+            });
+        })
+        .catch((error) => {
+            console.error("Error fetching data:", error);
+        });
+}
+
+function populateSelects() {
+    populateConcentrator();
+    populateNoiseMeter();
+}
 
 function set_contact_table(contactData) {
     var contactTable = new Tabulator("#contacts_table", {
@@ -113,6 +225,54 @@ function set_measurement_point_table(measurementPoint_data) {
             },
         ],
     });
+    measurementPointTable.on("rowClick", function (e, row) {
+        // window.location.href = "/measurement_point/" + row.getIndex();
+        console.log("point clicked");
+    });
+    measurementPointTable.on("rowSelectionChanged", function (data, rows) {
+        table_row_changed(data);
+    });
+    window.measurementPointTable = measurementPointTable;
+}
+
+function table_row_changed(data) {
+    if (data && data.length > 0) {
+        document.getElementById("editButton").disabled = false;
+        document.getElementById("deleteButton").disabled = false;
+        inputMeasurementPointId = data[0].id;
+        fetch_measurement_point_data(data[0]);
+    } else {
+        document.getElementById("editButton").disabled = true;
+        document.getElementById("deleteButton").disabled = true;
+    }
+}
+
+function fetch_measurement_point_data(data) {
+    noise_meter_data = [];
+    concentrator_data = [];
+    var pointName = document.getElementById("inputUpdatePointName");
+    var remarks = document.getElementById("inputUpdateRemarks");
+    var device_location = document.getElementById("inputUpdateDeviceLocation");
+
+    if (data) {
+        pointName.value = data.point_name;
+        remarks.value = data.remarks;
+        device_location.value = data.device_location;
+
+        concentrator_data.push({
+            concentrator_id: data.concentrator_id,
+            concentrator_label: data.concentrator_label,
+            device_id: data.device_id,
+        });
+
+        noise_meter_data.push({
+            noise_meter_id: data.noise_meter_id,
+            noise_meter_label: data.noise_meter_label,
+            serial_number: data.serial_number,
+        });
+
+        console.log(noise_meter_data);
+    }
 }
 
 function getProjectId() {
@@ -186,6 +346,91 @@ function create_users(projectId, csrfToken) {
             }
         });
     });
+}
+
+function handle_create_measurement_point() {
+    var csrfToken = document
+        .querySelector('meta[name="csrf-token"]')
+        .getAttribute("content");
+    var form = document.getElementById("measurement_point_create_form");
+
+    var formData = new FormData(form);
+
+    var formDataJson = {};
+    formData.forEach((value, key) => {
+        formDataJson[key] = value;
+    });
+
+    fetch(form.action, {
+        method: form.method,
+        headers: {
+            "X-CSRF-TOKEN": csrfToken,
+            "Content-Type": "application/json",
+            Accept: "application/json",
+            "X-Requested-With": "XMLHttpRequest",
+        },
+        body: JSON.stringify(formDataJson),
+    })
+        .then((response) => {
+            if (!response.ok) {
+                throw new Error(
+                    "Network response was not ok " + response.statusText
+                );
+            }
+            return response.json();
+        })
+        .then((json) => {
+            closeModal("measurementPointCreateModal");
+        })
+        .catch((error) => {
+            console.error("Error:", error);
+            alert("There was an error: " + error.message);
+        });
+    return false;
+}
+
+function handle_measurement_point_update() {
+    var csrfToken = document
+        .querySelector('meta[name="csrf-token"]')
+        .getAttribute("content");
+    var form = document.getElementById("measurementPointUpdateForm");
+
+    var formData = new FormData(form);
+
+    var formDataJson = {};
+    formData.forEach((value, key) => {
+        formDataJson[key] = value;
+    });
+
+    fetch(
+        "http://localhost:8000/measurement_points/" + inputMeasurementPointId,
+        {
+            method: "PATCH",
+            headers: {
+                "X-CSRF-TOKEN": csrfToken,
+                "Content-Type": "application/json",
+                Accept: "application/json",
+                "X-Requested-With": "XMLHttpRequest",
+            },
+            body: JSON.stringify(formDataJson),
+        }
+    )
+        .then((response) => {
+            if (!response.ok) {
+                throw new Error(
+                    "Network response was not ok " + response.statusText
+                );
+            }
+            return response.json();
+        })
+        .then((json) => {
+            closeModal("measurementPointUpdateModal");
+        })
+        .catch((error) => {
+            console.error("Error:", error);
+            alert("There was an error: " + error.message);
+        });
+    return false;
 }
 
 function handle_create_dummy_user() {
@@ -300,31 +545,46 @@ function deleteUser(event) {
         });
 }
 
-function handleDelete() {
+function handleDelete(event) {
+    if (event) {
+        event.preventDefault();
+    }
     var csrfToken = document
         .querySelector('meta[name="csrf-token"]')
         .getAttribute("content");
-    fetch("http://localhost:8000/measurement_points/" + inputprojectId, {
-        method: "DELETE",
-        headers: {
-            "X-CSRF-TOKEN": csrfToken,
-            Accept: "application/json",
-            "X-Requested-With": "XMLHttpRequest",
-        },
-    })
-        .then((response) => {
-            if (!response.ok) {
-                console.log("Error:", response);
-                throw new Error("Network response was not ok");
+
+    var confirmation = document.getElementById("inputDeleteConfirmation").value;
+    if (confirmation == "DELETE") {
+        fetch(
+            "http://localhost:8000/measurement_points/" +
+                inputMeasurementPointId,
+            {
+                method: "DELETE",
+                headers: {
+                    "X-CSRF-TOKEN": csrfToken,
+                    Accept: "application/json",
+                    "X-Requested-With": "XMLHttpRequest",
+                },
             }
-            return response.json();
-        })
-        .then((data) => {
-            console.log("Success:", data);
-        })
-        .catch((error) => {
-            console.error("Error:", error);
-        });
+        )
+            .then((response) => {
+                if (!response.ok) {
+                    console.log("Error:", response);
+                    throw new Error("Network response was not ok");
+                }
+                return response.json();
+            })
+            .then((data) => {
+                console.log("Success:", data);
+                closeModal("deleteConfirmationModal");
+            })
+            .catch((error) => {
+                console.error("Error:", error);
+            });
+    } else {
+        var error = document.getElementById("deleteConfirmationError");
+        error.hidden = false;
+    }
 }
 
 function handleUpdate() {
@@ -361,8 +621,6 @@ function handleUpdate() {
         .then((json) => {
             create_users(inputprojectId, csrfToken);
             closeModal("updateModal");
-            // location.reload() should only be called after successful operations
-            location.reload();
         })
         .catch((error) => {
             console.error("Error:", error);
@@ -379,10 +637,12 @@ function openModal(modalName) {
         userList = [];
         modalType = "update";
         populateUser("userUpdateSelectList");
-    } else if (modalName == "projectcreateModal") {
-        userList = [];
-        inputprojectId = "";
+    } else if (modalName == "measurementPointCreateModal") {
         modalType = "create";
+        populateSelects();
+    } else if (modalName == "measurementPointUpdateModal") {
+        modalType = "update";
+        populateSelects();
     }
 }
 
@@ -422,8 +682,11 @@ function closeModal(modal) {
     const modalElement = document.getElementById(modal);
     const modalInstance = bootstrap.Modal.getInstance(modalElement);
     modalInstance.hide();
+    location.reload();
 }
 
+window.handle_measurement_point_update = handle_measurement_point_update;
+window.handle_create_measurement_point = handle_create_measurement_point;
 window.handleDelete = handleDelete;
 window.handleUpdate = handleUpdate;
 window.openModal = openModal;
